Updated coding standards.
- Headers are now to use relative paths when including files within their libraries.
- Headers in project have been brought into conformance.

+++ include/Resource/Types.h
@@ -1,7 +1,7 @@
 #ifndef _RESOURCE_TYPES_H_
 #define _RESOURCE_TYPES_H_
 
-#include <Resource/Config.h>
+#include "Config.h"
 
 #include <cstdint>
 
@@ -12,7 +12,7 @@ namespace Resource
 
 	// Types to support bytestream metrics
 	typedef uint32_t seek_t;
-	typedef size_t bytesize_t;
+	typedef uint32_t bytesize_t;
 }
 
 #endif // _RESOURCE_TYPES_H_
